Updated installer

+++ installer.sh
@@ -1,10 +1,12 @@
-pip install discord --user
-pip install wikipedia --user
-#pip install google_auth_oauthlib --user
-#pip install google-api-python-client --user
-pip install you --user
-pip install pokedex.py --user
-pip install python-dotenv --user
-pip install --user currencyconverter
-touch .env
-printf "DISCORD_TOKEN=\nADMIN_ID=\n" > .env
+pip install discord --user || pip3 install discord --user
+pip install wikipedia --user || pip3 install wikipedia --user
+pip install you --user || pip3 install you --user
+pip install pokedex.py --user || pip3 install pokedex.py --user
+pip install python-dotenv --user || pip3 install python-dotenv --user
+pip install --user currencyconverter || pip3 install currencyconverter --user
+
+FILE=.env
+IF test -f "$FILE"; then
+    touch .env
+    printf "DISCORD_TOKEN=\nADMIN_ID=\n" > .env
+fi
